kooplearn_examples-master/examples/levy_driven_langevin/requirements.txt
einops==0.6.0
findiff==0.9.2
jax_md==0.2.5
jaxtyping==0.2.14
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
scienceplots==2.0.1
scipy==1.10.1
tqdm==4.65.0
tweepy==4.13.0
